refactor: 자동차 이름 생성 로직 클래스 분리
