+++ docs/README.md
@@ -1,15 +1,15 @@
 ## 기능 요구사항 정리
-* 자동차 이름 입력
-  * 이름은 5자 이하만 가능
-  * 구분자는 쉼표
-  * 잘못 입력 시 에러 문구를 출력하고 다시 입력받기
-* 몇 번 이동할 것인지 입력
-  * 숫자만 입력 가능
-  * 잘못 입력 시 에러 문구를 출력하고 다시 입력받기
-* 0-9 사이의 랜덤 숫자 생성하기
-* 자동차 이동
-  * 숫자가 4 이상이면 전진
-  * 숫자가 3 이하면 정지
+* [x] 자동차 이름 입력
+  * [x] 이름은 5자 이하만 가능
+  * [x] 구분자는 쉼표
+  * [x] 잘못 입력 시 에러 문구를 출력하고 다시 입력받기
+* [x] 몇 번 이동할 것인지 입력
+  * [x] 숫자만 입력 가능
+  * [x] 잘못 입력 시 에러 문구를 출력하고 다시 입력받기
+* [x] 0-9 사이의 랜덤 숫자 생성하기
+* [x] 자동차 이동
+  * [x] 숫자가 4 이상이면 전진
+  * [x] 숫자가 3 이하면 정지
 * 우승자 선정하기
   * 자동차 위치 값을 비교해서 가장 많이 전진한 차가 우승
   * 우승자는 여러명일 수 있음
